add starting direction toggle and readme
this change adds the README with instructions
on how to use the page and a few tips for how
to use the pattern based on the silver reed lk150.
it probably works with other flatbed knitting
machines too, but i don't know anything about
those unfortunately :p

while i was writing the README i also figured
it might be nice to add a starting direction
toggle since i thought maybe not everyone
starts their colorwork from the same direction
as i do, and i don't think it really matters.
you could just fix the direction on your own,
but that seems annoying when i could just add
the toggle myself for convenience.